refactor: move node/group/edge code into utilities
